move file into folder
